\name{as.character-methods}
\docType{methods}
\alias{as.character-methods}
\alias{as.character,Criteria-method}
\alias{as.character,CriteriaManager-method}
\alias{as.character,CriteriaSet-method}
\alias{as.character,CutoffManager-method}
\alias{as.character,TopNManager-method}
\title{Methods for Function as.character in Package `Wfccm'}
\description{Character Vectors}
\section{Methods}{
\describe{

\item{object = "Criteria"}{ ~~}

\item{object = "CriteriaManager"}{ ~~}

\item{object = "CriteriaSet"}{ ~~}

\item{object = "CutoffManager"}{ ~~}

\item{object = "TopNManager"}{ ~~}
}}
\keyword{methods}
