\name{as.Criteria-methods}
\docType{methods}
\alias{as.Criteria-methods}
\alias{as.Criteria,ANY-method}
\title{Methods for Function `as.Criteria' in Package `Wfccm'}
\description{
  Convert other types to \code{"Criteria"}.
}
\section{Methods}{
\describe{

\item{x = "ANY"}{Generic function}
}}
\seealso{
  \code{\link{as.Criteria}}
}
\keyword{methods}
