\name{c-methods}
\docType{methods}
\alias{c-methods}
\alias{c,Criteria,missing-method}
\title{Methods for Function c in Package `Wfccm'}
\description{
Methods for function \code{c} in Package `Wfccm'
}
\section{Methods}{
\describe{

\item{x = "Criteria", recursive = "missing"}{ ~~describe this method here }
}}
\keyword{methods}
