\name{coerce-methods}
\docType{methods}
\alias{coerce-methods}
\alias{coerce,Criteria,character-method}
\alias{coerce,character,Criteria-method}
\alias{coerce,CriteriaManager,character-method}
\alias{coerce,CriteriaSet,character-method}
\alias{coerce,CutoffManager,character-method}
\alias{coerce,TopNManager,character-method}
\title{Methods for Function coerce in Package `Wfccm'}
\description{
Methods for function \code{coerce} in Package `Wfccm'
}
\section{Methods}{
\describe{

\item{from = "Criteria", to = "character"}{ ~~describe this method here }

\item{from = "character", to = "Criteria"}{ ~~describe this method here }

\item{from = "CriteriaManager", to = "character"}{ ~~describe this method here }

\item{from = "CriteriaSet", to = "character"}{ ~~describe this method here }

\item{from = "CutoffManager", to = "character"}{ ~~describe this method here }

\item{from = "TopNManager", to = "character"}{ ~~describe this method here }
}}
\keyword{methods}
