\name{Extract-methods}
\docType{methods}
\alias{[-methods}
\alias{[<--methods}
\alias{[,Criteria,character,missing-method}
\alias{[,Criteria,character,numeric-method}
\alias{[,Criteria,logical,missing-method}
\alias{[,Criteria,logical,numeric-method}
\alias{[,Criteria,numeric,missing-method}
\alias{[,Criteria,numeric,numeric-method}
\alias{[,Criteria,missing,missing-method}
\alias{[,Criteria,missing,numeric-method}
\alias{[<-,Criteria,character,missing,Criteria-method}
\alias{[<-,Criteria,character,numeric,numeric-method}
\alias{[<-,Criteria,logical,missing,Criteria-method}
\alias{[<-,Criteria,logical,numeric,numeric-method}
\alias{[<-,Criteria,numeric,missing,Criteria-method}
\alias{[<-,Criteria,numeric,numeric,numeric-method}
\alias{[<-,Criteria,missing,missing,Criteria-method}
\alias{[<-,Criteria,missing,numeric,numeric-method}
\alias{[[-methods}
\alias{[[,Criteria,numeric,missing-method}
\alias{[[,CriteriaManager,ANY,ANY-method}
\alias{[[,CutoffManager,numeric,missing-method}
\alias{[[,TopNManager,numeric,missing-method}
\title{Methods to Extract or Replace in Package `Wfccm'}
\description{
Operators acting on vector types in Package `Wfccm' to extract or replace parts.
}
\section{Methods}{
\describe{

\item{x = "Criteria", i = "character", j = "missing"}{ ~~describe this method here }

\item{x = "Criteria", i = "character", j = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "logical", j = "missing"}{ ~~describe this method here }

\item{x = "Criteria", i = "logical", j = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "numeric", j = "missing"}{ ~~describe this method here }

\item{x = "Criteria", i = "numeric", j = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "missing", j = "missing"}{ ~~describe this method here }

\item{x = "Criteria", i = "missing", j = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "character", j = "missing", value = "Criteria"}{ ~~describe this method here }

\item{x = "Criteria", i = "character", j = "numeric", value = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "logical", j = "missing", value = "Criteria"}{ ~~describe this method here }

\item{x = "Criteria", i = "logical", j = "numeric", value = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "numeric", j = "missing", value = "Criteria"}{ ~~describe this method here }

\item{x = "Criteria", i = "numeric", j = "numeric", value = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "missing", j = "missing", value = "Criteria"}{ ~~describe this method here }

\item{x = "Criteria", i = "missing", j = "numeric", value = "numeric"}{ ~~describe this method here }

\item{x = "Criteria", i = "numeric", j = "missing"}{ ~~describe this method here }

\item{x = "CriteriaManager", i = "ANY", j = "ANY"}{ ~~describe this method here }

\item{x = "CutoffManager", i = "numeric", j = "missing"}{ ~~describe this method here }

\item{x = "TopNManager", i = "numeric", j = "missing"}{ ~~describe this method here }
}}
\keyword{methods}
