\name{length-methods}
\docType{methods}
\alias{length-methods}
\alias{length<--methods}
\alias{length,Criteria-method}
\alias{length,CriteriaManager-method}
\alias{length<-,CriteriaManager,ANY-method}
\alias{length,CutoffManager-method}
\alias{length,TopNManager-method}
\alias{length<-,TopNManager,numeric-method}
\title{Methods for Function length in Package `Wfccm'}
\description{
Methods for function \code{length} in Package `Wfccm'
}
\section{Methods}{
\describe{

\item{x = "Criteria"}{ ~~describe this method here }

\item{x = "CriteriaManager"}{ ~~describe this method here }

\item{x = "CriteriaManager", value = "ANY"}{ ~~describe this method here }

\item{x = "CutoffManager"}{ ~~describe this method here }

\item{x = "TopNManager"}{ ~~describe this method here }

\item{x = "TopNManager", value = "numeric"}{ ~~describe this method here }
}}
\keyword{methods}
