\name{names-methods}
\docType{methods}
\alias{names-methods}
\alias{names,Criteria-method}
\alias{names,CriteriaManager-method}
\alias{names<-,CriteriaManager,character-method}
\alias{names,CriteriaSet-method}
\title{Methods for Function names in Package `Wfccm'}
\description{
Methods for function \code{names} in Package `Wfccm'
}
\section{Methods}{
\describe{

\item{x = "Criteria"}{ ~~describe this method here }

\item{x = "CriteriaManager"}{ ~~describe this method here }

\item{x = "CriteriaManager", value = "character"}{ ~~describe this method here }

\item{x = "CriteriaSet"}{ ~~describe this method here }
}}
\keyword{methods}
