\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,Criteria-method}
\alias{show,CriteriaManager-method}
\alias{show,CriteriaSet-method}
\alias{show,CutoffManager-method}
\alias{show,TopNManager-method}
\title{Methods for Function show in Package `Wfccm'}
\description{
Methods for function \code{show} in Package `Wfccm'
}
\section{Methods}{
\describe{

\item{object = "Criteria"}{ ~~describe this method here }

\item{object = "CriteriaManager"}{ ~~describe this method here }

\item{object = "CriteriaSet"}{ ~~describe this method here }

\item{object = "CutoffManager"}{ ~~describe this method here }

\item{object = "TopNManager"}{ ~~describe this method here }
}}
\keyword{methods}
