\name{sign.wfccm}
\alias{sign.wfccm}
\title{WFCCM Sign Data}
\description{
  Applies signes to data.frame variables for use in WFCCM.
}
\usage{
  sign.wfccm(data, lead, rest)
}
\arguments{
  \item{data}{numeric data.frame or matrix.}
  \item{lead}{numeric or character of the index or name of the variable to get the sign to apply to the \code{rest} variables.}
  \item{rest}{numeric or character of the indices or names of the variables to apply the \code{lead} sign to.}
}
\keyword{utilities}
