\docType{package}
\name{WfccmR-package}
\alias{WfccmR-package}
\title{Weighted Flexible Compound Covariate Method}
\description{
%TODO: Update this section
Yu Shyr's method for high-dimensional data analysis.
\url{http://www.vicc.org/biostatistics/wfccm/wfccm2.htm}
}
\author{
Will Gray\cr
Department of Biostatistics\cr
Vanderbilt University\cr
\email{w.gray@vanderbilt.edu}
}
\keyword{models}
\keyword{package}
