\name{wfccmr}
\alias{wfccmr}
\title{WFCCMR}
\description{
  Runs all steps for WFCCM.
}
\usage{
wfccmr(data, group, tests = tests.wfccm(data, group), model, testdata = NULL, testgroup = NULL, ...)
}
\arguments{
  \item{data}{numeric matrix or data.frame.}
  \item{group}{2-level factor of group classification.}
  \item{tests}{numeric matrix or data.frame of statistical test values and p-values.}
  \item{model}{object of class \code{\link{CriteriaManager}}.}
  \item{testdata}{numeric matrix or data.frame.  Optional.}
  \item{testgroup}{2-level factor.  Optional.}
  \item{\dots}{additional parameters.}
}
\details{
}
\value{
}
\keyword{utilities}
