\name{write.CriteriaManager}
\alias{write.CriteriaManager}
\alias{write.CutoffManager}
\alias{write.TopNManager}
\title{Write CriteriaManager objects}
\description{
  Write objects of type \code{"CriteriaManager"}.
}
\usage{
write.CriteriaManager(x, file)
write.CutoffManager(x, file)
write.TopNManager(x, file)
}
\arguments{
  \item{x}{the object to write to a file.}
  \item{file}{the name of the file which the object is to be written to.}
}
\seealso{\code{\link{read.CriteriaManager}}}
\keyword{file}
